[nan-591] move unit test to a specific unit test file
